app/requirements.txt
bcrypt==4.2.0
colorama==0.4.6
iniconfig==2.0.0
mysql-connector-python==9.0.0
packaging==24.1
pluggy==1.5.0
prettytable==3.11.0
pytest==8.3.3
python-dotenv==1.0.1
tabulate==0.9.0
wcwidth==0.2.13
